[Refactor] 수면기록 패키지 구조 변경
Ref : #24
